lightrag/api/requirements.txt
aiofiles
ascii_colors
fastapi
numpy
pipmaster
python-dotenv
python-multipart
tenacity
tiktoken
uvicorn
ollama
openai
jiter
distro
networkx
oracledb